Creating doc to hold my imho on healthcare cv
Adding and starting a readme on healthcare claims and CV19.  Covering my IMHO on where and why the overall healthcare system is not going to sustain widespread claim fraud.  Although its not impossible for fraud to manifest - the ability for fraud to become pervasive in a system with many places for claims to be interrogated and studied - means its more difficult for fraud to be rampant.